Better indexes for the blog + describe how we force blog view
Note that the better indexes for the blog require an up to date DWiki
source *and* the latest dwiki.css (which is part of the source update,
but you need to push it into your production area).

Although in theory how Wandering Thoughts 'forces' a blog view (well,
makes it the default) is something that you can deduce from existing
DWiki documentation, I decided that it should be covered explicitly.
Probably this means that DWiki needs better documentation.