Add proper previous/next entry links with the entry titles
For layout reasons this requires a bunch of hacks. To look right, the
previous/next links need to go into the content area of the main <table>,
which means that they cannot be generated through a simple override of
datecrumbs.tmpl. So we have to turn off the standard datecrumbs.tmpl
and then shim our own into a different place.

This commit also includes some cleanup and simplification for the
index page header. Bad me for bundling both in to the same commit,
yes, yes.